backend/app/requirements.txt
fastapi
uvicorn
pandas
spacy
sentence-transformers
faiss-cpu
scikit-learn
rapidfuzz
sqlalchemy
